fix(DFC-923): bugfix and refactor
