+++ packages/frontend-analytics/src/analytics/formTracker/formTrackerUtils/isExcludedType/isExcludedType.ts
@@ -1,0 +1,8 @@
+export const isExcludedType = (element: HTMLInputElement): boolean => {
+  return (
+    element.type === "hidden" ||
+    element.type === "fieldset" ||
+    element.type === "button" ||
+    element.type === "submit"
+  );
+};
